the Long type is modified by the Integer

+++ school-service/src/main/java/com/backfcdev/schoolservice/controller/SchoolController.java
@@ -32,7 +32,7 @@ public class SchoolController {
     }
 
     @GetMapping("/{id}")
-    ResponseEntity<School> findById(@PathVariable Long id){
+    ResponseEntity<School> findById(@PathVariable Integer id){
         return ResponseEntity.ok(schoolService.findById(id));
     }
